+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 
 ### Fixes
 - avoid archived, fresh chats #3053
+- Also resync UIDs in folders that are not configured #2289
 - treat "NO" IMAP response to MOVE and COPY commands as an error #3058
 - Fix a bug where messages in the Spam folder created contact requests #3015
 - Fix a bug where drafts disappeared after some days #3067